chore: remove machines from node termination and topology test
